+++ src/API/exports.jl
@@ -82,6 +82,7 @@ export pypos
 export pypow
 export pyprint
 export pyrange
+export pyrepl
 export pyrepr
 export pyrowlist
 export pyrshift